refactor cache for dict
